isolated_shopping/requirements.txt
Django==3.8.5
djangorestframework==2.2
django-filter==2.4.0
